Moves correct_immersed_tendencies for IncompressibleModel
